Update the Color module
