+++ CHANGELOG.md
@@ -2,39 +2,48 @@
 
 > **Tags:**
 >
-> * [New Feature]
-> * [Bug Fix]
-> * [Breaking Change]
-> * [Documentation]
-> * [Internal]
-> * [Polish]
-> * [Experimental]
+> - [New Feature]
+> - [Bug Fix]
+> - [Breaking Change]
+> - [Documentation]
+> - [Internal]
+> - [Polish]
+> - [Experimental]
 
 **Note**: Gaps between patch versions are faulty/broken releases. **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 1.0.0
+
+- **Breaking Change**
+  - upgrade to `fp-ts@2.x.x` and latest TypeScript (3.9.2) (@IMax153)
+  - remove `free-canvas`
+  - `Color`
+    - rename `color` module to `Color` (@IMax153)
+    - fix implementation of `toCss` (@IMax153)
+
 # 0.5.0
 
-* **Breaking Change**
-  * upgrade to `fp-ts@1.x.x` (@gcanti)
+- **Breaking Change**
+  - upgrade to `fp-ts@1.x.x` (@gcanti)
 
 # 0.4.0
 
-* **Breaking Change**
-  * upgrade to latest `fp-ts` (0.6) and latest TypeScript (2.6) (@gcanti)
+- **Breaking Change**
+  - upgrade to latest `fp-ts` (0.6) and latest TypeScript (2.6) (@gcanti)
 
 # 0.3.0
 
-* **Breaking Change**
-  * upgrade to latest `fp-ts` (0.5.1) and latest TypeScript (2.5.2) (@gcanti)
+- **Breaking Change**
+  - upgrade to latest `fp-ts` (0.5.1) and latest TypeScript (2.5.2) (@gcanti)
 
 # 0.2.0
 
-* **Breaking Change**
-  * upgrade to latest `fp-ts` (@gcanti)
+- **Breaking Change**
+  - upgrade to latest `fp-ts` (@gcanti)
 
 # 0.1.0
 
-* **Breaking Change**
-  * upgrade to latest `fp-ts` (`graphics-ts` APIs are not changed though) (@gcanti)
-  * drop `lib-jsnext` folder
+- **Breaking Change**
+  - upgrade to latest `fp-ts` (`graphics-ts` APIs are not changed though) (@gcanti)
+  - drop `lib-jsnext` folder
