+++ src/index.ts
@@ -1,10 +1,11 @@
 import * as canvas from './canvas'
-export { canvas }
-import * as freeCanvas from './free-canvas'
-export { freeCanvas }
-import * as color from './color'
-export { color }
+import * as color from './Color'
 import * as drawing from './drawing'
-export { drawing }
 import * as font from './font'
-export { font }
+
+export {
+  canvas,
+  color,
+  drawing,
+  font
+}
